# src/main/java/com/produtos/apirest/dto/ProdutoResposta.java
package com.produtos.apirest.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.ToString;

import java.math.BigDecimal;
import java.util.Date;

@Data
@NoArgsConstructor
@AllArgsConstructor
@ToString
@JsonIgnoreProperties(ignoreUnknown = true)
@JsonInclude(JsonInclude.Include.NON_DEFAULT)
public class ProdutoResposta {

    private long id;
    private String nome;
    private String lote;
    private BigDecimal quantidade;
    private BigDecimal valor;
    private Date lastModifiedDate;
}